aws-agentcore/agentcore-runtime/orchestrator-agent/pyproject.toml
[project]
name = "neo4j-orchestrator-agent"
version = "0.1.0"
description = "Multi-agent orchestrator for Neo4j MCP server with Maintenance and Operations specialists"
requires-python = ">=3.10"
dependencies = [
    "bedrock-agentcore",
    "bedrock-agentcore-starter-toolkit",
    # LangGraph for multi-agent orchestration
    "langgraph>=0.2.0",
    "langgraph-supervisor>=0.0.1",
    "langchain>=0.3.0",
    "langchain-aws>=0.2.0",
    "langchain-mcp-adapters>=0.2.0",
    "mcp>=1.3.0",
    "httpx>=0.28.0",
    "boto3>=1.42.0",
    "pyyaml>=6.0",
    # Observability - ADOT for automatic tracing to CloudWatch
    "aws-opentelemetry-distro>=0.10.0",
]
